+++ lib/simple_auth/railtie.rb
@@ -5,5 +5,12 @@ module SimpleAuth
     generators do
       require "simple_auth/generator"
     end
+
+    config.after_initialize do
+      ActiveSupport.on_load(:active_record) do
+        SimpleAuth::Session.record_not_found_exceptions <<
+          ActiveRecord::RecordNotFound
+      end
+    end
   end
 end