+++ lib/simple_auth/session.rb
@@ -2,6 +2,10 @@
 
 module SimpleAuth
   class Session
+    def self.record_not_found_exceptions
+      @record_not_found_exceptions ||= []
+    end
+
     def self.create(**kwargs)
       new(**kwargs)
     end
@@ -21,6 +25,8 @@ module SimpleAuth
       return unless record_id_from_session
 
       @record ||= GlobalID::Locator.locate(record_id_from_session)
+    rescue *Session.record_not_found_exceptions
+      nil
     end
 
     def record_key